Initialize learningautoplayer
The use of the Decorator Design Pattern will enable the generation of
weighted moves to be delegated. Another method would be to have a bank
of functions; however, there is the ability to delegate the name
containing responsibilities.

In terms of performance, it may be advantageous to add such a weighted
generation function.